Fly to vehicles on map
When a new line is selected fly to the vehicles on the map. This removes
the need to find where the vehicles are. I am doing this with an average
of all the vehicles which is pretty good, but not perfect.

I also added a .env file because I accidentally commited my API key...

Don't worry, I deleted that key!